Fix broken test cases
Upgrade to pyquery2 broke a number of test cases, this large commit
fixes almost all broken tests, and everything impacted by the upgrade.

As a side effect, all of the tests now use only local versions of
HTML files, so none of the tests themselves will break due to data format
changes in the future.

The rest of the test failures will be fixed in subsequent changes.