Add Coaches Poll rankings to NCAAF
The NCAAF rankings functionality currently supports two of the three ranking types available on sports-reference.com. This change adds support for the third ranking type, the USA Today Coaches Poll. This commit includes testing for the new feature, as well as corrects a few minor docstring errors in the rankings.py file.

Signed-off-by: Samuel Smith <smitsam5@oregonstate.edu>